refactor(testWrapper): 유틸 함수로 분리
